# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Karat Hive API",
    "version": "0.1.0",
    "description": "Generated from Nest controllers + ZodValidationPipe schemas (`npm run openapi:generate`). Zod remains the single source of request shape (NFR-030, AD-BE-14). First artefact is partial: paths/methods and request Zod schemas are covered; response presenters are not yet dual-exported as Zod. Pre-code catalogue: docs/API-Route-Inventory.md — clients should treat this file as the evolving contract."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Karat Hive access JWT (not a Firebase ID token)."
      }
    },
    "schemas": {
      "SuccessMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "type": "string"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "nextCursor"
        ]
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMeta"
          }
        },
        "required": [
          "meta"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/requests": {
      "get": {
        "operationId": "Admin_listRequests",
        "tags": [
          "Admin"
        ],
        "summary": "GET /v1/admin/requests",
        "description": "Handler Admin.listRequests. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/requests/{id}": {
      "get": {
        "operationId": "Admin_getRequest",
        "tags": [
          "Admin"
        ],
        "summary": "GET /v1/admin/requests/{id}",
        "description": "Handler Admin.getRequest. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/otp/request": {
      "post": {
        "operationId": "Auth_otpRequest",
        "tags": [
          "Auth"
        ],
        "summary": "POST /v1/auth/otp/request",
        "description": "Handler Auth.otpRequest. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mobileNumber": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{6,14}$"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "REGISTER_CUSTOMER",
                      "REGISTER_VENDOR",
                      "CHANGE_MOBILE"
                    ]
                  }
                },
                "required": [
                  "mobileNumber",
                  "purpose"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/requests": {
      "get": {
        "operationId": "Request_listMyRequests",
        "tags": [
          "Request"
        ],
        "summary": "GET /v1/me/requests",
        "description": "Handler Request.listMyRequests. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "FIND_ORNAMENT",
                "SELL_OLD_GOLD",
                "GOLD_COIN",
                "GOLD_BULLION"
              ]
            },
            "required": false,
            "name": "requestType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "BUY",
                "SELL"
              ]
            },
            "required": false,
            "name": "direction",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/requests": {
      "post": {
        "operationId": "Request_createDraft",
        "tags": [
          "Request"
        ],
        "summary": "POST /v1/requests",
        "description": "Handler Request.createDraft. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestType": {
                    "type": "string",
                    "enum": [
                      "FIND_ORNAMENT",
                      "SELL_OLD_GOLD",
                      "GOLD_COIN",
                      "GOLD_BULLION"
                    ]
                  },
                  "direction": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ]
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "regionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "weightGrams": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "weightIsApproximate": {
                    "type": "boolean"
                  },
                  "purityKarat": {
                    "type": "string",
                    "enum": [
                      "24K",
                      "22K",
                      "21K",
                      "18K",
                      "K24",
                      "K22",
                      "K21",
                      "K18"
                    ]
                  },
                  "ornamentType": {
                    "type": "string",
                    "enum": [
                      "RING",
                      "CHAIN",
                      "BANGLE",
                      "NECKLACE",
                      "EARRING",
                      "BRACELET",
                      "PENDANT",
                      "OTHER"
                    ]
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "NEW",
                      "LIKE_NEW",
                      "USED",
                      "DAMAGED"
                    ]
                  },
                  "denominationGrams": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "mintOrRefiner": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "budgetMin": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "budgetMax": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "budgetIsFlexible": {
                    "type": "boolean"
                  },
                  "gemstones": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  },
                  "mediaKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 5
                  }
                },
                "required": [
                  "requestType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}": {
      "get": {
        "operationId": "Request_getById",
        "tags": [
          "Request"
        ],
        "summary": "GET /v1/requests/{id}",
        "description": "Handler Request.getById. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "Request_update",
        "tags": [
          "Request"
        ],
        "summary": "PATCH /v1/requests/{id}",
        "description": "Handler Request.update. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestType": {
                    "type": "string",
                    "enum": [
                      "FIND_ORNAMENT",
                      "SELL_OLD_GOLD",
                      "GOLD_COIN",
                      "GOLD_BULLION"
                    ]
                  },
                  "direction": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ]
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "regionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "weightGrams": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "weightIsApproximate": {
                    "type": "boolean"
                  },
                  "purityKarat": {
                    "type": "string",
                    "enum": [
                      "24K",
                      "22K",
                      "21K",
                      "18K",
                      "K24",
                      "K22",
                      "K21",
                      "K18"
                    ]
                  },
                  "ornamentType": {
                    "type": "string",
                    "enum": [
                      "RING",
                      "CHAIN",
                      "BANGLE",
                      "NECKLACE",
                      "EARRING",
                      "BRACELET",
                      "PENDANT",
                      "OTHER"
                    ]
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "NEW",
                      "LIKE_NEW",
                      "USED",
                      "DAMAGED"
                    ]
                  },
                  "denominationGrams": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "mintOrRefiner": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "budgetMin": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "budgetMax": {
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "exclusiveMinimum": true
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "budgetIsFlexible": {
                    "type": "boolean"
                  },
                  "gemstones": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  },
                  "mediaKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}/cancel": {
      "post": {
        "operationId": "Request_cancel",
        "tags": [
          "Request"
        ],
        "summary": "POST /v1/requests/{id}/cancel",
        "description": "Handler Request.cancel. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}/duplicate": {
      "post": {
        "operationId": "Request_duplicate",
        "tags": [
          "Request"
        ],
        "summary": "POST /v1/requests/{id}/duplicate",
        "description": "Handler Request.duplicate. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  }
}